introduced _2016_05_06_starting_point.docm for future reference
This document was the starting point, as writers of v3.0.0 had left it.
I included it here, in order to help compare the resulting document of
the 1st revision, with respect to the original one.

During development and review of the document, some changes were
accepted, and some were rejected.
It was impossible to keep every change and comment in the document.
It would make review and reasoning impossible.
So, each in each commit of the document I tried to leave only the
open changes and comments that still needed to be discussed, accepted, or
rejected.

One can trace additions and removals of each step, by inspecting the
commit history, in the repository