Issue #6: Bolts vs Screws. <nut> is optional
document is updated with a clear description of how screws differ from
bolts.
This is taken from the 'Machinery Handbook' referred in wiki,

As a result of this, the <nut> is not required under <bolt>,
because bolt always implies the existence of a nut.
If we have no dimensions of a nut, <nut> may be omitted.

* Changed description in Section 7.5.1
* <nut>, in Table 56 is now optional
* Figure 23 in Table 7.5.6.1 is now a <screw> and not a bolt without a
nut !